指标管理

+++ src/main/java/top/joylife/tracker/common/bean/dto/TransferCountDto.java
@@ -2,6 +2,8 @@ package top.joylife.tracker.common.bean.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TransferCountDto {
 
@@ -13,10 +15,10 @@ public class TransferCountDto {
     /**
      * 总收入
      */
-    private Integer sumEarnings;
+    private BigDecimal sumEarnings;
 
     /**
      * 总支出
      */
-    private Integer sumPayout;
+    private BigDecimal sumPayout;
 }
